Don't run external codan tool on excluded files
There should probably be a UI setting for this.  If you have a source
file excluded, you probably don't want to analyze it and might not even
have a proper environment to analyze it anyway.  For example, I exclude
all files not in the compilation database and run Clang-tidy on all of
the project. Skipping excluded files also means we don't try to run the
external tool on files that are not even c/c++ source files.

Change-Id: Ic63e852e30d2cf03e7a572874320621c4216abe6